requirements.txt
telethon
psycopg2-binary
python-dotenv
pillow
pytesseract
